comment-service - GlobalExceptionHandler.java   - 로깅을위해 @Slf4j 어노테이션 추가   - ResponseEntity반환타입을 String으로하여 상세한내용을 숨기고 커스텀메세지를 제공   - UnauthorizedException예외 추가 - Comment.java   - idx_user_uuid 인덱스 추가   - idx_post_uuid 인덱스 추가   - createdDate 컬럼에 updatable = false로 하여 변경되지 않게 설정 - OutboxEvent.java   - idx_saga_processed, idx_event_saga 인덱스 추가   - createdDate 컬럼에 updatable = false로 하여 변경되지 않게 설정 - CommentRepository.java   - findByPostUuidAndUserUuid 삭제 - CommentService.java   - commentUuid와 userUuid를 기준으로 데이터를 찾는것을 명확한 응답을 제공하기위해 분리 - application.properties   - spring.jpa.hibernate.ddl-auto=validate로 변경 - ErrorMessage.java   - 하드코딩하던 에러메세지를 열거형으로 선언 - UnauthorizedException.java   - RuntimeException을 확장하여 UnauthorizedException 클래스 생성

+++ comment-service/src/main/java/com/miniblog/comment/service/comment/CommentService.java
@@ -4,15 +4,16 @@ import com.miniblog.comment.dto.request.CommentCreatedRequestDTO;
 import com.miniblog.comment.dto.response.CommentResponseDTO;
 import com.miniblog.comment.dto.request.CommentUpdatedRequestDTO;
 import com.miniblog.comment.exception.NotFoundException;
+import com.miniblog.comment.exception.UnauthorizedException;
 import com.miniblog.comment.mapper.CommentMapper;
 import com.miniblog.comment.model.Comment;
 import com.miniblog.comment.repository.comment.CommentRepository;
 import com.miniblog.comment.service.outbox.OutboxEventService;
 import com.miniblog.comment.util.ProducedEventType;
-import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.UUID;
 
@@ -25,9 +26,7 @@ public class CommentService {
     private final OutboxEventService outboxEventService;
 
     @Transactional
-    public CommentResponseDTO createComment(
-            String userUuid,
-            CommentCreatedRequestDTO commentCreatedRequestDTO) {
+    public CommentResponseDTO createComment(String userUuid, CommentCreatedRequestDTO commentCreatedRequestDTO) {
         UUID userUuidAsUUID = UUID.fromString(userUuid);
 
         Comment comment = commentMapper.createToEntity(userUuidAsUUID, commentCreatedRequestDTO);
@@ -38,15 +37,16 @@ public class CommentService {
     }
 
     @Transactional
-    public CommentResponseDTO updateComment(
-            String userUuid,
-            String commentUuid,
-            CommentUpdatedRequestDTO commentUpdatedRequestDTO) {
+    public CommentResponseDTO updateComment(String userUuid, String commentUuid, CommentUpdatedRequestDTO commentUpdatedRequestDTO) {
         UUID userUuidAsUUID = UUID.fromString(userUuid);
         UUID commentUuidAsUUID = UUID.fromString(commentUuid);
-
-        Comment comment = commentRepository.findByCommentUuidAndUserUuid(commentUuidAsUUID, userUuidAsUUID)
-                .orElseThrow(() -> new NotFoundException("Comment not found or user not authorized: commentUuid = " + commentUuidAsUUID + ", userUuid = " + userUuidAsUUID));
+        // 댓글 조회
+        Comment comment = commentRepository.findById(commentUuidAsUUID)
+                .orElseThrow(() -> new NotFoundException("댓글을 찾을 수 없습니다: commentUuid = " + commentUuid));
+        // 권한 확인
+        if (!comment.getUserUuid().equals(userUuidAsUUID)) {
+            throw new UnauthorizedException("이 댓글을 수정할 권한이 없습니다: userUuid = " + userUuid);
+        }
         commentMapper.updateToEntity(comment, commentUpdatedRequestDTO);
         commentRepository.save(comment);
         outboxEventService.createOutboxEvent(comment, ProducedEventType.COMMENT_UPDATED);
@@ -55,14 +55,16 @@ public class CommentService {
     }
 
     @Transactional
-    public void deleteComment(
-            String userUuid,
-            String postUuid) {
+    public void deleteComment(String userUuid, String commentUuid) {
         UUID userUuidAsUUID = UUID.fromString(userUuid);
-        UUID commentUuidAsUUID = UUID.fromString(postUuid);
-
-        Comment comment = commentRepository.findByCommentUuidAndUserUuid(commentUuidAsUUID, userUuidAsUUID)
-                .orElseThrow(() -> new NotFoundException("Comment not found or user not authorized: commentUuid = " + commentUuidAsUUID + ", userUuid = " + userUuidAsUUID));
+        UUID commentUuidAsUUID = UUID.fromString(commentUuid);
+        // 댓글 조회
+        Comment comment = commentRepository.findById(commentUuidAsUUID)
+                .orElseThrow(() -> new NotFoundException("댓글을 찾을 수 없습니다: commentUuid = " + commentUuid));
+        // 권한 확인
+        if (!comment.getUserUuid().equals(userUuidAsUUID)) {
+            throw new UnauthorizedException("이 댓글을 삭제할 권한이 없습니다: userUuid = " + userUuid);
+        }
         commentRepository.delete(comment);
 
         outboxEventService.createOutboxEvent(comment, ProducedEventType.COMMENT_DELETED);
